Switch to single module project
