+++ rest-api/pom.xml
@@ -20,9 +20,9 @@
 
     <dependencies>
         <dependency>
-            <groupId>io.openexchange</groupId>
-            <artifactId>domain-model</artifactId>
-            <version>${project.parent.version}</version>
+            <groupId>org.projectlombok</groupId>
+            <artifactId>lombok</artifactId>
+            <optional>true</optional>
         </dependency>
         <dependency>
             <groupId>org.springframework.boot</groupId>
