+++ rest-api/src/main/java/io/openexchange/openexchange/ocp/s2i/java/model/HelloWorld.java
