+++ rest-api/src/test/java/io/openexchange/openexchange/ocp/s2i/java/model/HomeWorldTest.java
@@ -1,3 +1,5 @@
+package io.openexchange.openexchange.ocp.s2i.java.model;
+
 import org.junit.jupiter.api.Test;
 
 public class HomeWorldTest {
